pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.11.0",
    "flask>=3.1.1",
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.1.0",
    "tweepy>=4.15.0",
]
